Create module boundaries for `pane_grid` logic
