+++ native/src/widget/pane_grid/pane.rs
@@ -1,0 +1,2 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub struct Pane(pub(super) usize);
